reciever_lambda/requirements.txt
boto3==1.28.20
botocore==1.31.20
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
jmespath==1.0.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
